### 2021-11-01 #3874 @whedon
[PRE REVIEW]: Annoroad Browser: A multi-omic visualization tool
**Submitting author:** @Spartanzhao (<a href="http://orcid.org/0000-0001-5293-820X">Yue Zhao</a>)
**Repository:** <a href="https://github.com/Spartanzhao/Annoroad-Browser/" target ="_blank">https://github.com/Spartanzhao/Annoroad-Browser/</a>
**Version:** v.1.0.0
