### 2021-11-01 #3874 @danielskatz
👋 @Spartanzhao - thanks for your submission.  Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.

Also, please note that your paper does not compile, with an error in the bib file. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). My guess is that the "/" or " " or both in the bib entry name is causing the problem - you might change this to "AB-compartment".  You probably also will need to change "WashU Browser" to "WashU-Browser" or similar, then also change the .md file where these are cited.

Please feel free to make changes to your .bib or .md file, then use the command `@whedon generate pdf` to make a new PDF. Whedon commands need to be the first entry in a new comment.
