### 2021-11-03 #3874 @Spartanzhao
Yes, I will fix this problems as soon as possible. My work is trying to build a tool which is much better than WashU browser and pyGenomeTrack, and in some situations, replace IGV. This tool is very easy to expand, other people can easily add any new functions if they want, or customize their heat map color series. This is a very useful multi-omic visualization tool. We have many papers are going to publish based on this tool. Because all the Hi-C, epigenome analysis all require this tool's help in our research center.
