### 2021-11-14 #3874 @Spartanzhao
> @Spartanzhao have you fixed the issue with your paper yet?

I think that I had fixed the issue on reference. But I am not sure whether there are some other issues.
