### 2021-11-14 #3874 @Spartanzhao
> @Spartanzhao have you fixed the issue with your paper yet?

@jmschrei , I have fixed the problem, my github have passed the test on https://whedon.theoj.org/
