### 2021-11-20 #3916 @pkerpedjiev
So just to be clear, comments on the paper should be communicated by opening an issue in the software's issues page?
