### 2021-11-20 #3916 @pkerpedjiev
Perhaps I'm missing something but I don't see any screenshots in either the [README](https://github.com/Spartanzhao/Annoroad-Browser) or the [paper itself](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03916/joss.03916/10.21105.joss.03916.pdf).

The authors should spend a bit of time to clean up the README. The one sample command I could find was near the bottom, not properly formatted as a code block and yielded an error when trying to run it:

```
(base) toc2~/projects/Annoroad-Browser(main|✔) % python bin/Annoroad_Lego_Browser.py -fl show.list -o MLLT3_4C.png -chr chr9 -st 20620542 -ed 20625542 -b1 1000000 -b2 15000000
python: can't open file 'bin/Annoroad_Lego_Browser.py': [Errno 2] No such file or directory
```
