### 2021-12-01 Spartanzhao/Annoroad-Browser#1 @elliothershberg
Change name of tool
Hi @Spartanzhao , I am one of the reviewers for your JOSS submission for Annoroad. This looks like a helpful package for making visualizations. However, I'm not sure that it seems accurately named. For the most well known genome browsers such as [UCSC](http://genome.ucsc.edu/), [WashU Epigenome Browser](https://epigenomegateway.wustl.edu/), [IGV](https://igv.org/), and [JBrowse](https://jbrowse.org/jb2/), these are all interactive web applications for browsing genomic data. I'm not sure that this package constitutes what mosts genomics researchers would consider to be a browser.

Would you consider changing the name of the package to simply Annoroad, or Annoroad-Plot, Annoroad-Viz, Annoroad-Graphics, etc.?

cc @pkerpedjiev
