### 2022-01-05 #3916 @jmschrei
I think he means that he has opened issues in the GitHub repo for you to address, e.g. https://github.com/Spartanzhao/Annoroad-Browser/issues/1
