### 2022-01-07 Spartanzhao/Annoroad-Browser#1 @Spartanzhao
Okay, No problem,  Shall I rename my paper's title too? Am I have to withdraw my paper and resubmit again? Or just change the repository name and PAPER.md in my repository?
