### 2022-01-09 Spartanzhao/Annoroad-Browser#1 @elliothershberg
Cool, thanks for doing that! I assume you don't need to resubmit based on this and we can probably keep going with the review.
