### 2022-04-30 #3916 @pkerpedjiev
Sorry, delayed response but is there a way to reject the submission?

I'm still not entirely sure what to review. The README is cluttered and missing the most important sections of a README: installation and running. @Spartanzhao added an `install.readme.txt` which defies all convention for Python packages (installation information should be in the README). After installing the package using the suggested pip command, I have no idea how to run it. The pip installed package does not install `Annoroad_Lego_Browser.py` as an executable.

I don't mean to sound rude, but if this was a package that I had a use for, I'd take one look at the README and look for an alternative. And I would find an excellent one in [pyGenomeTracks](https://github.com/deeptools/pyGenomeTracks), which seems to offer the same functionality in a usable package.
