### 2022-05-02 #3916 @danielskatz
@Spartanzhao - We encourage you to improve this software, and potentially resubmit an improved version. If you do so, please note this issue in your resubmission.  Alternatively, please see https://joss.readthedocs.io/en/latest/submitting.html#other-venues-for-reviewing-and-publishing-software-packages for other suggestions for how you might receive credit for your work.

@pkerpedjiev, @elliothershberg - thanks for your help in the initial review work and feedback to the author and editor, and thanks to @jmschrei for agreeing to edit, even if this didn't come to a successful conclusions
