### 2022-05-07 #3916 @Spartanzhao
@[pkerpedjiev](https://github.com/pkerpedjiev) my Annoroad_Lego_Browser is absolutely much better and more handy than pyGenomeTrack. pyGenomeTrack is absolutely not an excellent package.
